refactor: MainActivity 책임 간소화
MainActivity의 게임 재개, 도메인 로직과 UI 연결, 예외 발생시 재시도 책임을 다른 객체로 위임

- GameResume : 게임 재개 역할
- OmokBoardAdapter : 도메인 로직과 UI 연결 역할
- HandlingExceptionsUtils: 예외 발생시 재시도 역할
