+++ app/src/main/java/woowacourse/omok/utils/HandlingExceptionUtils.kt
@@ -1,0 +1,17 @@
+package woowacourse.omok.utils
+
+import android.widget.ImageView
+import com.google.android.material.snackbar.Snackbar
+
+object HandlingExceptionUtils {
+    fun <T> retryUntilSuccess(
+        view: ImageView,
+        action: () -> T,
+    ): T? =
+        runCatching {
+            action()
+        }.getOrElse {
+            Snackbar.make(view, it.localizedMessage, Snackbar.LENGTH_LONG).show()
+            return null
+        }
+}
